Finish all data types
