+++ plutarch-ledger-api/plutarch-ledger-api.cabal
@@ -51,6 +51,15 @@ library
     Plutarch.LedgerApi.Utils
     Plutarch.LedgerApi.Value
 
+  other-modules:
+    Plutarch.LedgerApi.Address
+    Plutarch.LedgerApi.Contexts
+    Plutarch.LedgerApi.Credential
+    Plutarch.LedgerApi.Crypto
+    Plutarch.LedgerApi.Scripts
+    Plutarch.LedgerApi.Time
+    Plutarch.LedgerApi.Tx
+
   build-depends:
     , bytestring
     , cryptonite
