+++ plutarch-ledger-api/src/Plutarch/LedgerApi/Crypto.hs
@@ -1,0 +1,64 @@
+{-# OPTIONS_GHC -Wno-orphans #-}
+
+-- Mirrors the equivalent V1 module in plutus-ledger-api
+module Plutarch.LedgerApi.Crypto (
+  PPubKeyHash (..),
+) where
+
+import Plutarch.LedgerApi.Utils (Mret)
+import Plutarch.Lift (
+  DerivePConstantViaBuiltin (DerivePConstantViaBuiltin),
+  PConstantDecl,
+  PUnsafeLiftDecl (PLifted),
+ )
+import Plutarch.Prelude
+import Plutarch.TryFrom (PTryFrom (PTryFromExcess, ptryFrom'))
+import Plutarch.Unsafe (punsafeCoerce)
+import PlutusLedgerApi.V3 qualified as Plutus
+
+-- | @since 2.0.0
+newtype PPubKeyHash (s :: S) = PPubKeyHash (Term s PByteString)
+  deriving stock
+    ( -- | @since 2.0.0
+      Generic
+    )
+  deriving anyclass
+    ( -- | @since 2.0.0
+      PlutusType
+    , -- | @since 2.0.0
+      PIsData
+    , -- | @since 2.0.0
+      PEq
+    , -- | @since 2.0.0
+      PPartialOrd
+    , -- | @since 2.0.0
+      POrd
+    , -- | @since 2.0.0
+      PShow
+    )
+
+-- | @since 2.0.0
+instance DerivePlutusType PPubKeyHash where
+  type DPTStrat _ = PlutusTypeNewtype
+
+-- | @since 2.0.0
+instance PUnsafeLiftDecl PPubKeyHash where
+  type PLifted PPubKeyHash = Plutus.PubKeyHash
+
+-- | @since 2.0.0
+deriving via
+  (DerivePConstantViaBuiltin Plutus.PubKeyHash PPubKeyHash PByteString)
+  instance
+    PConstantDecl Plutus.PubKeyHash
+
+-- | @since 2.0.0
+instance PTryFrom PData (PAsData PPubKeyHash) where
+  type PTryFromExcess PData (PAsData PPubKeyHash) = Mret PPubKeyHash
+  ptryFrom' opq = runTermCont $ do
+    unwrapped <- tcont . plet $ ptryFrom @(PAsData PByteString) opq snd
+    tcont $ \f ->
+      pif
+        (plengthBS # unwrapped #== 28)
+        (f ())
+        (ptraceInfoError "ptryFrom(PPubKeyHash): must be 28 bytes long")
+    pure (punsafeCoerce opq, pcon . PPubKeyHash $ unwrapped)
