+++ plutarch-ledger-api/src/Plutarch/LedgerApi/Tx.hs
@@ -1,0 +1,211 @@
+{-# OPTIONS_GHC -Wno-orphans #-}
+
+-- Mirrors the equivalent V3 module in plutus-ledger-api, with some V2 stuff as
+-- well
+module Plutarch.LedgerApi.Tx (
+  PTxId (..),
+  PTxOutRef (..),
+  PTxOut (..),
+  POutputDatum (..),
+) where
+
+import Plutarch.DataRepr (
+  DerivePConstantViaData (DerivePConstantViaData),
+  PDataFields,
+ )
+import Plutarch.LedgerApi.Address (PAddress)
+import Plutarch.LedgerApi.AssocMap qualified as AssocMap
+import Plutarch.LedgerApi.Scripts (PDatum, PDatumHash, PScriptHash)
+import Plutarch.LedgerApi.Utils (Mret, PMaybeData)
+import Plutarch.LedgerApi.Value qualified as Value
+import Plutarch.Lift (
+  DerivePConstantViaBuiltin (DerivePConstantViaBuiltin),
+  PConstantDecl,
+  PUnsafeLiftDecl (PLifted),
+ )
+import Plutarch.Prelude
+import Plutarch.TryFrom (PTryFrom (PTryFromExcess, ptryFrom'))
+import Plutarch.Unsafe (punsafeCoerce)
+import PlutusLedgerApi.V3 qualified as Plutus
+
+{- | Hashed with @BLAKE2b-256@.
+
+@since 3.1.0
+-}
+newtype PTxId (s :: S) = PTxId (Term s PByteString)
+  deriving stock
+    ( -- | @since 2.0.0
+      Generic
+    )
+  deriving anyclass
+    ( -- | @since 2.0.0
+      PlutusType
+    , -- | @since 2.0.0
+      PIsData
+    , -- | @since 2.0.0
+      PEq
+    , -- | @since 2.0.0
+      PPartialOrd
+    , -- | @since 2.0.0
+      POrd
+    , -- | @since 2.0.0
+      PShow
+    )
+
+-- | @since 3.1.0
+instance DerivePlutusType PTxId where
+  type DPTStrat _ = PlutusTypeNewtype
+
+-- | @since 2.0.0
+instance PUnsafeLiftDecl PTxId where
+  type PLifted PTxId = Plutus.TxId
+
+-- | @since 3.1.0
+deriving via
+  (DerivePConstantViaBuiltin Plutus.TxId PTxId PByteString)
+  instance
+    PConstantDecl Plutus.TxId
+
+-- | @since 3.1.0
+instance PTryFrom PData (PAsData PTxId) where
+  type PTryFromExcess PData (PAsData PTxId) = Mret PTxId
+  ptryFrom' opq = runTermCont $ do
+    unwrapped <- tcont . plet $ ptryFrom @(PAsData PByteString) opq snd
+    tcont $ \f ->
+      pif
+        (plengthBS # unwrapped #== 32)
+        (f ())
+        (ptraceInfoError "ptryFrom(PTxId): must be 32 bytes long")
+    pure (punsafeCoerce opq, pcon . PTxId $ unwrapped)
+
+{- | Reference to a transaction output, with an index referencing which exact
+output we mean.
+
+@since 2.0.0
+-}
+newtype PTxOutRef (s :: S)
+  = PTxOutRef
+      ( Term
+          s
+          ( PDataRecord
+              '[ "id" ':= PTxId
+               , "idx" ':= PInteger
+               ]
+          )
+      )
+  deriving stock
+    ( -- | @since 2.0.0
+      Generic
+    )
+  deriving anyclass
+    ( -- | @since 2.0.0
+      PlutusType
+    , -- | @since 2.0.0
+      PIsData
+    , -- | @since 2.0.0
+      PDataFields
+    , -- | @since 2.0.0
+      PEq
+    , -- | @since 2.0.0
+      PPartialOrd
+    , -- | @since 2.0.0
+      POrd
+    , -- | @since 2.0.0
+      PTryFrom PData
+    , -- | @since 2.0.0
+      PShow
+    )
+
+-- | @since 2.0.0
+instance DerivePlutusType PTxOutRef where
+  type DPTStrat _ = PlutusTypeData
+
+-- | @since 2.0.0
+instance PUnsafeLiftDecl PTxOutRef where
+  type PLifted PTxOutRef = Plutus.TxOutRef
+
+-- | @since 2.0.0
+deriving via
+  (DerivePConstantViaData Plutus.TxOutRef PTxOutRef)
+  instance
+    PConstantDecl Plutus.TxOutRef
+
+-- | @since 2.0.0
+newtype PTxOut (s :: S)
+  = PTxOut
+      ( Term
+          s
+          ( PDataRecord
+              '[ "address" ':= PAddress
+               , "value" ':= Value.PValue 'AssocMap.Sorted 'Value.Positive
+               , "datum" ':= POutputDatum
+               , "referenceScript" ':= PMaybeData PScriptHash
+               ]
+          )
+      )
+  deriving stock
+    ( -- | @since 2.0.0
+      Generic
+    )
+  deriving anyclass
+    ( -- | @since 2.0.0
+      PlutusType
+    , -- | @since 2.0.0
+      PIsData
+    , -- | @since 2.0.0
+      PDataFields
+    , -- | @since 2.0.0
+      PEq
+    , -- | @since 2.0.0
+      PShow
+    )
+
+-- | @since 2.0.0
+instance DerivePlutusType PTxOut where
+  type DPTStrat _ = PlutusTypeData
+
+-- | @since 2.0.0
+instance PUnsafeLiftDecl PTxOut where
+  type PLifted PTxOut = Plutus.TxOut
+
+-- | @since 2.0.0
+deriving via
+  (DerivePConstantViaData Plutus.TxOut PTxOut)
+  instance
+    PConstantDecl Plutus.TxOut
+
+-- | @since 2.0.0
+data POutputDatum (s :: S)
+  = PNoOutputDatum (Term s (PDataRecord '[]))
+  | POutputDatumHash (Term s (PDataRecord '["datumHash" ':= PDatumHash]))
+  | -- | Inline datum as per
+    -- [CIP-0032](https://github.com/cardano-foundation/CIPs/blob/master/CIP-0032/README.md)
+    POutputDatum (Term s (PDataRecord '["outputDatum" ':= PDatum]))
+  deriving stock
+    ( -- | @since 2.0.0
+      Generic
+    )
+  deriving anyclass
+    ( -- | @since 2.0.0
+      PlutusType
+    , -- | @since 2.0.0
+      PIsData
+    , -- | @since 2.0.0
+      PEq
+    , -- | @since 2.0.0
+      PShow
+    )
+
+-- | @since 2.0.0
+instance DerivePlutusType POutputDatum where
+  type DPTStrat _ = PlutusTypeData
+
+-- | @since 2.0.0
+instance PUnsafeLiftDecl POutputDatum where
+  type PLifted POutputDatum = Plutus.OutputDatum
+
+-- | @since 2.0.0
+deriving via
+  (DerivePConstantViaData Plutus.OutputDatum POutputDatum)
+  instance
+    PConstantDecl Plutus.OutputDatum
